Move all files to a single getHelpContent
