+++ src/components/SidePane/HelpComponents/HelpContent.tsx
@@ -6,7 +6,7 @@ import {ScrollView} from 'react-native-gesture-handler';
 import {useOnyx} from 'react-native-onyx';
 import HeaderGap from '@components/HeaderGap';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
-import getHelpContent from '@components/SidePane/HelpContent/getHelpContent';
+import getHelpContent from '@components/SidePane/getHelpContent';
 import useEnvironment from '@hooks/useEnvironment';
 import useLocalize from '@hooks/useLocalize';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
